% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/StaticLayer.R
\docType{methods}
\name{AddLabels}
\alias{AddLabels}
\alias{AddLabels,StaticLayer,StaticLayer-method}
\alias{AddLabels,StaticLayer-method}
\title{Method AddLabels}
\usage{
AddLabels(object)

\S4method{AddLabels}{StaticLayer}(object)
}
\arguments{
\item{object}{Object of class StaticLayer}
}
\description{
Method AddLabels

#
}

