% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MapLayout.R
\docType{methods}
\name{AddLayout}
\alias{AddLayout}
\alias{AddLayout,MapLayout-method}
\title{Method AddLayout}
\usage{
AddLayout(object)

\S4method{AddLayout}{MapLayout}(object)
}
\arguments{
\item{object}{Object of class MapLayout}
}
\description{
Method AddLayout
}

