% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/borders.R
\name{borders}
\alias{borders}
\title{polygons to borders}
\usage{
borders(spdf, tol = 1, id = NULL)
}
\arguments{
\item{spdf}{SpatialPolygonsDataFrame}

\item{tol}{tolerance to detect contiguity (in map units)}

\item{id}{unique id on the spdf}
}
\value{
SpatialLinesDataFrame
}
\description{
between polygons
}
\details{
Extract borders from a SpatialPolygonsDataFrame
}
\examples{
contig <-borders(spdf=FRdep.spdf)
plot(FRdep.spdf,lwd=0.2)
plot(contig,col="red",lwd=0.4,add=TRUE)
}

