% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/choroLayer.R, R/internL_choro.R
\name{choro}
\alias{choro}
\alias{choroLayer}
\title{choro}
\usage{
choroLayer(spdf, df, spdfid = NULL, dfid = NULL, var, distr = NULL,
  col = NULL, nbclass = NULL, method = "quantile", pal1 = "blue.pal",
  pal2 = "green.pal", breakval = NULL, alpha = FALSE, add = T)

choro(spdf, df, spdfid = NULL, dfid = NULL, var, distr = NULL,
  col = NULL, nbclass = NULL, method = "quantile", pal1 = "blue.pal",
  pal2 = "green.pal", breakval = NULL, alpha = FALSE)
}
\arguments{
\item{spdf}{spatial dataframe (geometries)}

\item{df}{dataframe (data)}

\item{spdfid}{id of the spdf (if null, the first column is used)}

\item{dfid}{id of the df (if null, the first column is used)}

\item{var}{field used}

\item{distr}{vector of classes}

\item{col}{vector of colors}

\item{nbclass}{number of classes targeted (if null, the Huntsberger method is used)}

\item{method}{discretization method ("sd", "equal", "quantile", "jenks","q6","geom")}

\item{pal1}{color palette 1}

\item{pal2}{color palette 2}

\item{breakval}{Diverging value}

\item{alpha}{alphaeffect enhancing contrasts with opacity}

\item{add}{add=T}

\item{method}{methods of discretisation : "sd", "equal", "quantile", "jenks", "geom", "q6"}

\item{spdf}{spatial dataframe (geometries)}

\item{df}{dataframe (data)}

\item{spdfid}{id of the spdf (if null, the first column is used)}

\item{dfid}{id of the df (if null, the first column is used)}

\item{var}{field used}

\item{distr}{vector of classes}

\item{col}{vector of colors}

\item{nbclass}{number of classes targeted (if null, the Huntsberger method is used)}

\item{method}{discretization method ("sd", "equal", "quantile", "jenks","q6","geom")}

\item{pal1}{color palette 1}

\item{pal2}{color palette 2}

\item{breakval}{Diverging value}

\item{alpha}{alphaeffect enhancing contrasts with opacity}

\item{method}{methods of discretisation : "sd", "equal", "quantile", "jenks", "geom", "q6"}
}
\value{
plot

list
}
\description{
plot a chorpoleth layer

add color gradients to spdf according to data classes
}
\details{
plot a chorpoleth layer

xxx
}
\examples{
choroLayer(spdf = TNdeleg.spdf, df = TNdeleg, var =  "housing", nbclass = 20, pal1="red.pal",pal2="green.pal",breakval=10000,method = "equal",add=FALSE,alpha=TRUE)
# xxx
}

