% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/choroLayer.R
\name{choroLayer}
\alias{choroLayer}
\title{choroLayer}
\usage{
choroLayer(spdf, df, spdfid = NULL, dfid = NULL, var, distr = NULL,
  col = NULL, nbclass = NULL, method = "quantile", pal1 = "blue.pal",
  pal2 = "green.pal", breakval = NULL, alpha = FALSE, add = T)
}
\arguments{
\item{spdf}{spatial dataframe (geometries)}

\item{df}{dataframe (data)}

\item{spdfid}{id of the spdf (if null, the first column is used)}

\item{dfid}{id of the df (if null, the first column is used)}

\item{var}{field used}

\item{distr}{vector of classes}

\item{col}{vector of colors}

\item{nbclass}{number of classes targeted (if null,
the Huntsberger method is used)}

\item{method}{discretization method ("sd", "equal",
"quantile", "jenks","q6","geom")}

\item{pal1}{color palette 1}

\item{pal2}{color palette 2}

\item{breakval}{Diverging value}

\item{alpha}{alphaeffect enhancing contrasts with opacity}

\item{add}{add=T}
}
\value{
plot
}
\description{
plot a chorpoleth layer
}
\details{
plot a chorpoleth layer
}
\examples{
choroLayer(spdf = TNdeleg.spdf, df = TNdeleg, var =  "housing", nbclass = 20,
pal1="red.pal",pal2="green.pal",breakval=10000,method = "equal",add=FALSE,
alpha=TRUE)
}

