% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/internL_palette.R
\name{display.gradient}
\alias{display.gradient}
\title{display.gradiant}
\usage{
display.gradient(v)
}
\arguments{
\item{v}{vector with hex colors}
}
\value{
plot
}
\description{
This function displays one vector with hex colors
}
\details{
xxx
}
\examples{
myGradient <- palette("green.pal",8,"blue.pal",5,middle=FALSE,alphaeffect=FALSE)
display.gradient(myGradient)
}

