% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/internL_palette.R
\name{display.palette.all}
\alias{display.palette.all}
\title{display.palette.all}
\usage{
display.palette.all(nb = 10)
}
\arguments{
\item{nb}{numeric number of colors by palettes (from 1 to 20)}
}
\value{
vector
}
\description{
This function displays all the palettes available in the package
}
\details{
xxx
}
\examples{
display.palette.all(1)
display.palette.all(5)
display.palette.all(8)
display.palette.all(12)
display.palette.all(20)
}

