% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/internL_palette.R
\name{display.palette}
\alias{display.palette}
\title{display.palette}
\usage{
display.palette(name)
}
\arguments{
\item{name}{character name of the palette available in the package}
}
\value{
plot
}
\description{
This function displays all the content of one palette
}
\details{
xxx
}
\examples{
display.palette("orange.pal")
display.palette("sand.pal")
}

