% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{freeCarto}
\alias{freeCarto}
\alias{freeCarto-package}
\title{freeCarto Package}
\description{
This package contains various functions.
\itemize{
\item{\code{\link{layoutLayer}} to make a nice frame around a map}
\item{\code{\link{staticLayer}} to add layers on a map}
\item{\code{\link{labelLayer}} to add labels on a map}
}
}

