% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MapLayout.R
\name{LayoutMap}
\alias{LayoutMap}
\title{LayoutMap function.}
\usage{
LayoutMap(title = "Title of the map, year", sources = "Source(s)",
  author = "Author(s)", col = "black", txtcol = "white", scale = 0,
  frame = TRUE, north = FALSE)
}
\arguments{
\item{title}{Title of the map}

\item{sources}{Sources of the map}

\item{author}{Author of the map}

\item{col}{Color of the frame}

\item{txtcol}{Color of the Title}

\item{scale}{Size of the scale}

\item{frame}{Whether displaying a frame or not}

\item{north}{Noth arrow}
}
\description{
LayoutMap function.
}
\details{
Must only be called after a plot
}
\examples{
data("TNdeleg")
StaticMap(obj = TNdeleg.spdf)
LabelMap(obj = TNdeleg.spdf, data = TNdeleg, cex = 0.4)
LayoutMap(title = "Hell Yeah!", sources = "Sources Inconnues",
         author = "Mister T",scale = 150, frame = FALSE, north = TRUE )
}

