% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/OSMLayer.R
\name{OSMMap}
\alias{OSMMap}
\title{OSMMap function.}
\usage{
OSMMap(obj, type = "osm", zoom, add = FALSE)
}
\arguments{
\item{obj}{Spatial*DataFrame}

\item{type}{Type of map used in the \code{OpenStreetMap} package}

\item{zoom}{Zoom level of the map}

\item{add}{Whether to add the layer to an existing map (TRUE) or not (FALSE)}
}
\description{
OSMMap function.
}
\examples{
data("TNdeleg")
# OSMMap(obj = TNdeleg.spdf, type = "osm", add = FALSE)
StaticMap(obj = TNdeleg.spdf, add = TRUE, col = NA)
LayoutMap(title = "Hell Yeah!", sources = "Sources Inconnues, OSM",
          author = "Mister T",scale = 150, frame = FALSE, north = TRUE )
}

