% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/internL_palette.R
\name{palette}
\alias{palette}
\title{palette}
\usage{
palette(pal1, n1, pal2 = NULL, n2 = NULL, middle = FALSE,
  alphaeffect = FALSE)
}
\arguments{
\item{pal1}{name of the color gradiant}

\item{n1}{number of colors (up to 20)}

\item{pal2}{name of a second color gradiant (optionnal)}

\item{n2}{number of colors for the second color gradiant (up to 20) (optionnal)}

\item{middle}{boolean. if true, add a neutral color in the middle (optionnal)}

\item{alphaeffect}{boolean. If true, enhance contrasts by adding opacity (optionnal)}
}
\value{
vector
}
\description{
This function call and combine colors
}
\details{
This function call and combine colors from the colors object with several paraemeters
}
\examples{
# Simple gradient in blue
palette("blue.pal",20)

# double gradient blue / red
palette(pal1="blue.pal",n1=10, pal2="red.pal",n2=10)

# Adding a neutral color
palette(pal1="blue.pal",n1=10,pal2="red.pal",n2=10,middle=TRUE)

# Enhancing contrasts with alphaeffect
palette(pal1="blue.pal",n1=10,pal2="red.pal",n2=10,middle=TRUE,alphaeffect=TRUE)

# The double gradient can be asymetric
palette(pal1="blue.pal",n1=5,pal2="red.pal",n2=15,middle=TRUE,alphaeffect=TRUE)
}

