% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/propSymbolsLayer.R
\name{propSymbolsLayer}
\alias{propSymbolsLayer}
\title{Proportional Symbols Layer}
\usage{
propSymbolsLayer(spdf, df, spdfid = NA, dfid = NA, var,
  symbols = "circles", col = "#E84923", col2 = "#7DC437", breakval = 0,
  k = 0.2, fixmax = FALSE, pos = "bottomleft", title = var,
  add = TRUE)
}
\arguments{
\item{spdf}{Spatial*DataFrame}

\item{df}{DataFrame with Ids and Labels}

\item{spdfid}{Ids of the obj Spatial*DataFrame}

\item{dfid}{Ids of the DataFrame}

\item{var}{Symbols variable}

\item{symbols}{Type of symbol ("circles", "squares", "height")}

\item{col}{Symbols color}

\item{col2}{Symbols color if the break value (\code{breakval})is relevant}

\item{breakval}{Breaking value if 2 colors are needed}

\item{k}{Share of the map occupied by symbols}

\item{fixmax}{Whether the maximum value is fixed or not}

\item{pos}{Position of the legend}

\item{title}{Title of the legend}

\item{add}{Whether to add the layer to an existing map (TRUE) or not (FALSE)}
}
\description{
Proportional Symbols Layer
}
\examples{
data("TNdeleg")
}

