% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/staticLayers.R
\name{staticLayer}
\alias{staticLayer}
\title{staticLayer}
\usage{
staticLayer(spdf, col = "#EFDEC1", border = "black", lwd = 1, cex = 1,
  pch = 19, add = FALSE)
}
\arguments{
\item{spdf}{Spatial*DataFrame}

\item{col}{Filling color}

\item{border}{Border color}

\item{lwd}{Border thickness}

\item{cex}{cex}

\item{pch}{pch}

\item{add}{Whether to add the layer to an existing map (TRUE) or not (FALSE)}
}
\description{
ddd
}
\details{
Just a wrapper for sp::plot()
}
\examples{
data("TNdeleg")
}

