% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/internL_choro.R
\name{choro}
\alias{choro}
\title{choro}
\usage{
choro(spdf, df, spdfid = NULL, dfid = NULL, var, distr = NULL,
  col = NULL, nbclass = NULL, method = "quantile", pal1 = "blue.pal",
  pal2 = "green.pal", breakval = NULL, alpha = FALSE)
}
\arguments{
\item{spdf}{spatial dataframe (geometries)}

\item{df}{dataframe (data)}

\item{spdfid}{id of the spdf (if null, the first column is used)}

\item{dfid}{id of the df (if null, the first column is used)}

\item{var}{field used}

\item{distr}{vector of classes}

\item{col}{vector of colors}

\item{nbclass}{number of classes targeted (if null, the Huntsberger method is used)}

\item{method}{discretization method ("sd", "equal", "quantile", "jenks","q6","geom")}

\item{pal1}{color palette 1}

\item{pal2}{color palette 2}

\item{breakval}{Diverging value}

\item{alpha}{alphaeffect enhancing contrasts with opacity}
}
\value{
list
}
\description{
add color gradients to spdf according to data classes
}
\details{
xxx
}
\examples{
# xxx
}

